Move load_yaml_file to common
