+++ benchmarks/dynamo/torchbench.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 
-import functools
 import gc
 import importlib
 import logging
@@ -11,15 +10,13 @@ import warnings
 from collections import namedtuple
 from os.path import abspath, exists
 
-import yaml
-
 import torch
 
 
 try:
-    from .common import BenchmarkRunner, main
+    from .common import BenchmarkRunner, load_yaml_file, main
 except ImportError:
-    from common import BenchmarkRunner, main
+    from common import BenchmarkRunner, load_yaml_file, main
 
 from torch._dynamo.testing import collect_results, reduce_to_scalar_loss
 from torch._dynamo.utils import clone_inputs
@@ -71,37 +68,6 @@ def setup_torchbench_cwd():
     return original_dir
 
 
-@functools.lru_cache(maxsize=1)
-def load_yaml_file():
-    filename = "torchbench.yaml"
-    filepath = os.path.join(os.path.dirname(__file__), filename)
-
-    with open(filepath) as f:
-        data = yaml.safe_load(f)
-
-    internal_file_path = os.path.join(os.path.dirname(__file__), "fb", filename)
-    if os.path.exists(internal_file_path):
-        with open(internal_file_path) as f:
-            internal_data = yaml.safe_load(f)
-            data.update(internal_data)
-
-    def flatten(lst):
-        for item in lst:
-            if isinstance(item, list):
-                yield from flatten(item)
-            else:
-                yield item
-
-    def maybe_list_to_set(obj):
-        if isinstance(obj, dict):
-            return {k: maybe_list_to_set(v) for k, v in obj.items()}
-        if isinstance(obj, list):
-            return set(flatten(obj))
-        return obj
-
-    return maybe_list_to_set(data)
-
-
 def process_hf_reformer_output(out):
     assert isinstance(out, list)
     # second output is unstable
@@ -134,7 +100,7 @@ class TorchBenchmarkRunner(BenchmarkRunner):
 
     @property
     def _config(self):
-        return load_yaml_file()
+        return load_yaml_file("torchbench.yaml")
 
     @property
     def _skip(self):
